Add Windows PowerShell Version
Converted NMS-Decompressinator to PowerShell.

This script searches the registry for both Gog and Steam Installation Paths to ascertain the installation path for No Man's Sky
It then uses this information to present the user with a choice as to which path they wish the script to act upon (Gog or Steam).

The main advantage is to mitigate the requirement of placing the script and tools into the game directory to perform its functions.
This is no longer required in this version as long as this script and the tools reside within the same directory, it will work since this
script accounts for its location and where the tools reside.

It also includes some basic error checking in order to report to the user if certain conditions are not met. If no installations for the
game are found. Or if the tools (psarc.exe  / NMSResign.exe) are not found in the script directory. 

I also took the liberty of styling the output to try and give a consistent interface when the script is operating or reporting errors to
the user. This is purely aesthetic and doesn't interfere with the functionality.

It has been tested multiple times with both the Gog and Steam versions of the game and functions as expected.